Bug 993204 - Operation params 'MessageId' and 'DestinationQueue' are swapped in HornetQ "Move Message" Operation UI.
JMSResourceComponent has a generic implementation of invokeOperation.
In this method, it was expected that when reading the Configuration object parameters would come in the same order as in the plugin descriptor.

But in a Configuration object properties are stored in a map so order is not guaranteed.

It's now expected that:
* when an operation has a single param, name attribute has the following form "[type:]name"
* when an operation has multiple params, name attribute has the following form "pos:[type:]name"